Update timestep tensor type in SDXLPipeline from float16 to float32 for improved precision and mark subproject commit in pipeline/compel as dirty to indicate uncommitted changes.

+++ pipeline/pipeline.py
@@ -43,7 +43,7 @@ class SDXLPipeline:
         for t in timesteps:
             latent_model_input = latents_torch
 
-            timestep_numpy = np.array([t.item()], dtype=np.float16)
+            timestep_numpy = np.array([t.item()], dtype=np.float32)
 
             noise_pred_np = self.unet(
                 latent_model_input.cpu().numpy(),
